feat: introduce margin order
This patch introduces margin orders, i.e. the mean to define a _market_ order based on the margin to be used and not the quantity as usual. The quantity is then being calculated once it's matched against one or more limit orders.

This patch also introduces fractional quantities. The better solution would have been to make the leverage a function of quantity and margin but that seemed to be a bigger task. Hence, fractional quantities it is.

The biggest challenge is that coordinator and trader agree on the rounding of the quantity if it can be fractional. I've defined it as 2 decimal places. If they are not aligned, we will at closing the position because the coordinator believe we want to resize the position.

+++ mobile/lib/common/amount_text.dart
@@ -40,5 +40,5 @@ String formatSats(Amount amount) {
 
 String formatUsd(Usd usd) {
   final formatter = NumberFormat("\$ #,###,###,###,###", "en");
-  return formatter.format(usd.usd);
+  return formatter.format(usd.usd.toDouble());
 }
